Restructure questions building logic
