+++ src/final-mbz-builder.js
@@ -1,15 +1,29 @@
-const path = require('path');
-const { buildGroupsXml } = require('./json-to-mbz/groups');
-const { buildCoursesXml } = require('./json-to-mbz/course/course');
+const fs = require("fs");
+const path = require("path");
+const { buildGroupsXml } = require("./json-to-mbz/groups");
+const { buildCoursesXml } = require("./json-to-mbz/course/course");
+const { buildQuestionsXml } = require("./json-to-mbz/questions");
 
-const finalDir = path.join(__dirname, '..', 'final-mbz');
-const groupsJsonFilePath = "./exported_data/json/export-file-groups-2024-07-01-11-30-19.json";
-const courseJsonFilePath = "./exported_data/json/export-file-sfwd-courses-2024-07-01-11-30-19.json";
+const finalDir = path.join(__dirname, "..", "final-mbz");
 
-function createFinalMoodleBackup(finalDir, groupsJsonFilePath, courseJsonFilePath) {
-    buildGroupsXml(groupsJsonFilePath, finalDir);
-    buildCoursesXml(courseJsonFilePath, finalDir);
+const outputDir = `./output/${finalDir}`;
+
+// Ensure the output directory exists
+if (!fs.existsSync(outputDir)) {
+  fs.mkdirSync(outputDir, { recursive: true });
 }
 
+const groupsJsonFilePath =
+  "./exported_data/json/export-file-groups-2024-07-01-11-30-19.json";
+const courseJsonFilePath =
+  "./exported_data/json/export-file-sfwd-courses-2024-07-01-11-30-19.json";
+const questionsJsonFilePath =
+  "./exported_data/json/export-file-quiz_pro_1-2024-07-01-11-30-19.json";
 
-createFinalMoodleBackup(finalDir, groupsJsonFilePath, courseJsonFilePath);
+function createFinalMoodleBackup() {
+  buildGroupsXml(groupsJsonFilePath, finalDir);
+  buildCoursesXml(courseJsonFilePath, finalDir);
+  buildQuestionsXml(questionsJsonFilePath, finalDir);
+}
+
+createFinalMoodleBackup();
