+++ src/mbz-backup-builder.js
@@ -33,9 +33,6 @@ const {
 } = require("./components/course/blocks/generateBlocksFiles");
 const { generateGroupsXml } = require("./components/generateGroupsXml");
 
-const { createFilePath } = require("./utils/utils");
-const config = require("../config.json");
-
 function createMoodleBackup(outputDir) {
   // Create subdirectories
   const backupDirs = ["course"];
@@ -61,7 +58,7 @@ function createMoodleBackup(outputDir) {
   generateGroupsXml(outputDir);
   generateMoodleBackup(outputDir);
   generateOutcomesXml(outputDir);
-  generateQuestionsXml(createFilePath(config, "quizPath"), outputDir);
+  generateQuestionsXml(outputDir);
   generateRolesXml(outputDir);
   generateScalesXml(outputDir);
